api/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "synapse-api"
version = "0.1.0"
description = "Synapse backend — Django 5 + Ninja + JWT + Postgres + Temporal client"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2,<5.3",
    "django-ninja>=1.3,<2",
    "django-ninja-jwt>=5.3",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.0",
    "dj-database-url>=2.2",
    "gunicorn>=22",
    "email-validator>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "model-bakery>=1.18",
    "freezegun>=1.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["synapse*", "apps*"]
exclude = ["tests*", ".venv*"]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [".venv", "migrations", "**/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "synapse.settings.test"
addopts = "-q --reuse-db"
testpaths = ["tests", "apps"]
python_files = ["test_*.py", "tests.py"]
